# rs/generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "RESTful API文档",
    "version" : "v1.0",
    "title" : "RESTful API文档",
    "termsOfService" : "http://www.showtime.com/",
    "contact" : {
      "name" : "showtime",
      "url" : "http://www.showtime.com",
      "email" : "594470882@qq.com"
    }
  },
  "basePath" : "/rs",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/api/v1/banners/{id}" : {
      "get" : {
        "tags" : [ "轮播图接口" ],
        "summary" : "获取单个轮播图",
        "description" : "通过id获取轮播图",
        "operationId" : "getBanner",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "轮播图id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful request"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      },
      "put" : {
        "tags" : [ "轮播图接口" ],
        "summary" : "更新轮播图",
        "description" : "更新轮播图信息",
        "operationId" : "updateUser",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "轮播图id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "轮播图信息",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/BannerView"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful"
          },
          "404" : {
            "description" : "not found"
          },
          "409" : {
            "description" : "conflict"
          },
          "500" : {
            "description" : "internal Server Error"
          }
        }
      },
      "delete" : {
        "tags" : [ "轮播图接口" ],
        "summary" : "删除轮播图",
        "description" : "通过id删除轮播图",
        "operationId" : "deleteBanner",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "轮播图id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful request"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      }
    },
...
    "/api/v1/sysusers" : {
      "get" : {
        "tags" : [ "后台管理用户接口" ],
        "summary" : "获取后台管理用户列表",
        "description" : "通过查询条件获取后台管理用户列表",
        "operationId" : "getSysUsers",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "userName",
          "in" : "query",
          "description" : "用户名",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "enabled",
          "in" : "query",
          "description" : "状态",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "pageNumber",
          "in" : "query",
          "description" : "页数",
          "required" : false,
          "type" : "integer",
          "default" : 0,
          "format" : "int32"
        }, {
          "name" : "pageSize",
          "in" : "query",
          "description" : "每页加载量",
          "required" : false,
          "type" : "integer",
          "default" : 10,
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful request"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      },
      "post" : {
        "tags" : [ "后台管理用户接口" ],
        "summary" : "创建后台管理用户",
        "description" : "创建一个后台管理用户",
        "operationId" : "createSysUser",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "后台管理用户",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/SysUserView"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "successful operation",
            "headers" : {
              "location" : {
                "type" : "string",
                "description" : "URL of new created resource"
              }
            }
          },
          "409" : {
            "description" : "conflict"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      },
      "delete" : {
        "tags" : [ "后台管理用户接口" ],
        "summary" : "删除后台管理用户",
        "description" : "批量删除后台管理用户",
        "operationId" : "deleteSysUsers",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "后台管理用户ids",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful request"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      }
    },
    "/api/v1/sysusers/username/{username}" : {
      "get" : {
        "tags" : [ "后台管理用户接口" ],
        "summary" : "获取后台管理用户",
        "description" : "通过id获取后台管理用户",
        "operationId" : "getSysUserByUserName",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "username",
          "in" : "path",
          "description" : "后台管理用户userName",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful request"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      }
    },
    "/api/v1/sysusers/{id}" : {
      "get" : {
        "tags" : [ "后台管理用户接口" ],
        "summary" : "获取后台管理用户",
        "description" : "通过id获取后台管理用户",
        "operationId" : "getSysUser",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "后台管理用户id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful request"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      },
      "put" : {
        "tags" : [ "后台管理用户接口" ],
        "summary" : "更新后台管理用户",
        "description" : "更新后台管理用户信息",
        "operationId" : "updateUser",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "用户id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "用户信息",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/SysUserView"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful"
          },
          "404" : {
            "description" : "not found"
          },
          "409" : {
            "description" : "conflict"
          },
          "500" : {
            "description" : "internal Server Error"
          }
        }
      },
      "delete" : {
        "tags" : [ "后台管理用户接口" ],
        "summary" : "删除后台管理用户",
        "description" : "通过id删除后台管理用户",
        "operationId" : "deleteSysUser",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "后台管理用户id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful request"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      }
    },
...
  },
  "definitions" : {
    "BannerView" : {
      "type" : "object",
      "required" : [ "createTime", "updateTime" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "createTime" : {
          "type" : "integer",
          "format" : "int64"
        },
        "updateTime" : {
          "type" : "integer",
          "format" : "int64"
        },
        "enabled" : {
          "type" : "integer",
          "format" : "int32"
        },
        "auditors" : {
          "type" : "string"
        },
        "auditStatus" : {
          "type" : "integer",
          "format" : "int32"
        },
        "auditContent" : {
          "type" : "string"
        },
        "auditTime" : {
          "type" : "integer",
          "format" : "int64"
        },
        "title" : {
          "type" : "string"
        },
        "content" : {
          "type" : "string"
        },
        "subTitle" : {
          "type" : "string"
        },
        "imageUrl" : {
          "type" : "string"
        },
        "redirectUrl" : {
          "type" : "string"
        },
        "sequence" : {
          "type" : "integer",
          "format" : "int32"
        },
        "search" : {
          "type" : "string"
        }
      }
    },
...
    "SysAuthority" : {
      "type" : "object",
      "required" : [ "createTime", "updateTime" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "createTime" : {
          "type" : "integer",
          "format" : "int64"
        },
        "updateTime" : {
          "type" : "integer",
          "format" : "int64"
        },
        "enabled" : {
          "type" : "integer",
          "format" : "int32"
        },
        "auditors" : {
          "type" : "string"
        },
        "auditStatus" : {
          "type" : "integer",
          "format" : "int32"
        },
        "auditContent" : {
          "type" : "string"
        },
        "auditTime" : {
          "type" : "integer",
          "format" : "int64"
        },
        "hasAuthority" : {
          "type" : "boolean",
          "default" : false
        },
        "name" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        }
      }
    },
...
    "SysRole" : {
      "type" : "object",
      "required" : [ "createTime", "updateTime" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "createTime" : {
          "type" : "integer",
          "format" : "int64"
        },
        "updateTime" : {
          "type" : "integer",
          "format" : "int64"
        },
        "enabled" : {
          "type" : "integer",
          "format" : "int32"
        },
        "auditors" : {
          "type" : "string"
        },
        "auditStatus" : {
          "type" : "integer",
          "format" : "int32"
        },
        "auditContent" : {
          "type" : "string"
        },
        "auditTime" : {
          "type" : "integer",
          "format" : "int64"
        },
        "hasRole" : {
          "type" : "boolean",
          "default" : false
        },
        "englishName" : {
          "type" : "string"
        },
        "chineseName" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "sysAuthorities" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/SysAuthority"
          }
        }
      }
    },
...
    "SysUserView" : {
      "type" : "object",
      "required" : [ "createTime", "updateTime" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "createTime" : {
          "type" : "integer",
          "format" : "int64"
        },
        "updateTime" : {
          "type" : "integer",
          "format" : "int64"
        },
        "enabled" : {
          "type" : "integer",
          "format" : "int32"
        },
        "auditors" : {
          "type" : "string"
        },
        "auditStatus" : {
          "type" : "integer",
          "format" : "int32"
        },
        "auditContent" : {
          "type" : "string"
        },
        "auditTime" : {
          "type" : "integer",
          "format" : "int64"
        },
        "userName" : {
          "type" : "string"
        },
        "password" : {
          "type" : "string"
        },
        "sysRole" : {
          "$ref" : "#/definitions/SysRole"
        }
      }
    }
  }
}